backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
dj-database-url==2.1.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

# Testing dependencies
coverage==7.4.0
pytest>=7.0.0,<8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-asyncio==0.23.4
factory-boy==3.3.0